Remove wpilib.run
- Corresponding logic will be moved to robotpy-meta
- __main__ will no longer be the user's robot code, but I don't think
  anything uses that?
- Fixes #51
